requirements.txt
asgiref==3.9.1
attrs==25.3.0
colorama==0.4.6
coverage==7.10.6
dj-database-url==3.0.1
Django==5.2.5
django-filter==25.1
django-phonenumber-field==8.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
factory_boy==3.3.1
Faker==33.1.0
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
packaging==25.0
phonenumbers==9.0.12
pillow==11.3.0
pluggy==1.6.0
psycopg2-binary==2.9.10
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.3.4
pytest-cov==6.0.0
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.27.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
